Removed more C compiler warnings on Darwin.

+++ Native/inspector/darwin/darwinTeleNativeThread.c
@@ -55,17 +55,17 @@ Java_com_sun_max_tele_debug_darwin_DarwinTeleNativeThread_nativeReadRegisters(JN
     OsFloatingPointRegistersStruct osFloatRegisters;
     OsStateRegistersStruct osStateRegisters;
 
-    if (integerRegistersLength > sizeof(canonicalIntegerRegisters)) {
+    if (integerRegistersLength > (jint) sizeof(canonicalIntegerRegisters)) {
         log_println("buffer for integer register data is too large");
         return false;
     }
 
-    if (stateRegistersLength > sizeof(canonicalStateRegisters)) {
+    if (stateRegistersLength > (jint) sizeof(canonicalStateRegisters)) {
         log_println("buffer for state register data is too large");
         return false;
     }
 
-    if (floatingPointRegistersLength > sizeof(canonicalFloatingPointRegisters)) {
+    if (floatingPointRegistersLength > (jint) sizeof(canonicalFloatingPointRegisters)) {
         log_println("buffer for floating point register data is too large");
         return false;
     }
@@ -168,7 +168,7 @@ static jboolean suspendOtherThreads(jlong task, thread_t current) {
         log_println("thread_info() failed on thread to step");
         return false;
     }
-    for (j = 0; j < info.suspend_count; j++) {
+    for (j = 0; j < (unsigned) info.suspend_count; j++) {
         // unsuspend the current thread.
         thread_resume(current);
     }
@@ -192,7 +192,7 @@ static jboolean unsuspendOtherThreads(jlong task, thread_t current) {
                 log_println("thread_info() failed when single stepping");
                 return false;
             }
-            for (j = 0; j < info.suspend_count; j++) {
+            for (j = 0; j < (unsigned) info.suspend_count; j++) {
                 kret = thread_resume(thread_list[i]);
                 if (kret != KERN_SUCCESS) {
                     log_println("thread_resume() failed when single stepping");
